+++ controllers/contacts/index.js
@@ -1,0 +1,14 @@
+const getContacts =  require("./getContacts");
+const getContactId =  require("./getContactId");
+const deleteContact =  require("./deleteContact");
+const patchFavorite =  require("./patchFavorite");
+const putContact =  require("./putContact");
+
+
+module.exports = {
+    getContacts,
+    getContactId,
+    deleteContact,
+    patchFavorite,
+    putContact
+};
