+++ controllers/index.js
@@ -1,0 +1,16 @@
+const {
+  getContacts,
+  getContactId,
+  createContact,
+  deleteContact,
+  putContact,
+  patchFavorite,
+} = require("./contacts/index");
+module.exports = {
+  getContacts,
+  getContactId,
+  createContact,
+  deleteContact,
+  putContact,
+  patchFavorite,
+};
