+++ routes/api/contacts.js
@@ -1,21 +1,22 @@
 const express = require("express");
 // імпорт контролерра
-const ContactController = require("../../controllers/contact");
+// const ContactController = require("../../controllers/index");
+const { getContacts, getContactId, createContact, deleteContact, putContact, patchFavorite } = require("../../controllers/index");
 // маршрутизація
 const router = express.Router();
 // змінна для парсингу json
 const jsonParsers = express.json();
 // отримання всіх контактів
-router.get("/", ContactController.getContacts);
+router.get("/", getContacts);
 // отримання контакта по id
-router.get("/:contactId", ContactController.getContactId)
+router.get("/:contactId", getContactId)
 // додавання контакту
-router.post("/", jsonParsers, ContactController.creatreContact);
+router.post("/", jsonParsers, createContact);
 // видалення контакту
-router.delete("/:contactId", ContactController.deleteContact);
+router.delete("/:contactId", deleteContact);
 // оновлення даних в контакті
-router.put("/:contactId", jsonParsers, ContactController.putContact);
+router.put("/:contactId", jsonParsers, putContact);
 
-router.patch("/:contactId", jsonParsers, ContactController.patchFavorite)
+router.patch("/:contactId", jsonParsers, patchFavorite)
 
 module.exports = router;
